pyproject.toml
[tool.poetry]
name = "Sample-API-Data-Scraper"
version = "0.1.0"
description = ""
authors = ["Maksym Litvinenko <maksym.lithuanian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
motor = "^3.6.0"
pydantic = "^2.0"
python-dotenv = "^1.0.1"
scrapy = "^2.11.2"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
